+++ tests/test_scoring.py
@@ -4,10 +4,7 @@ import json
 
 sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
 
-from recursive_thinking_ai import (  # noqa: E402
-    EnhancedRecursiveThinkingChat,
-    CoRTConfig,
-)
+from core.chat import EnhancedRecursiveThinkingChat, CoRTConfig
 
 
 def test_score_called(monkeypatch):
